Exercise 7. Part 1 returning centroids as int

+++ ex7/findClosestCentroids.py
@@ -26,5 +26,5 @@ def findClosestCentroids(X, centroids):
 
 # =============================================================
 
-    return idx
+    return idx.astype('int')
